requirements.txt
pytest==7.1.1
requests==2.27.1
allure-pytest==2.9.45